Moved mechanism and tele-op classes
