+++ TeamCode/src/main/java/org/pattonvillerobotics/teleop/MainTeleop.java
@@ -1,8 +1,9 @@
-package org.pattonvillerobotics;
+package org.pattonvillerobotics.teleop;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
+import org.pattonvillerobotics.mechanisms.GlyphGrabber;
 import org.pattonvillerobotics.commoncode.opmodes.OpModeGroups;
 import org.pattonvillerobotics.commoncode.robotclasses.drive.SimpleMecanumDrive;
 import org.pattonvillerobotics.commoncode.robotclasses.gamepad.GamepadData;
